fixing test and linter errors
