+++ comp/metadata/hostsysteminfo/impl/hostsysteminfo_nix_test.go
@@ -1,0 +1,21 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build test && !windows
+
+package hostsysteminfoimpl
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNewSystemInfoProvider_EndUserDeviceMode(t *testing.T) {
+
+	hh := getTestHostSystemInfo(t, nil)
+	// Should be disabled for non-Windows platforms
+	assert.False(t, hh.InventoryPayload.Enabled, "Should not be enabled for non-Windows platforms")
+}
